Extract stopMediaStreamTrack into separate file and remove workaround fix which looks for that track in track controller
Also, bump version to 2.2.0 and move getMediaStreamTrackControllerInstance into mediaStreamTrack directory
