+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "media-stream-track-controller",
-  "version": "2.1.3",
+  "version": "2.2.0",
   "description": "Controller library for browser-based MediaStreamTrack objects",
   "main": "src/index.js",
   "scripts": {
